MPISphericalHarmonics

Documentation for MPISphericalHarmonics.

MPISphericalHarmonics.magneticField

MPISphericalHarmonics.magneticField — Method

magneticField(tDesign::SphericalTDesign, field::Union{AbstractArray{T,2},AbstractArray{T,3}}, x::Variable, y::Variable, z::Variable; L::Int=Int(tDesign.T/2), calcSolid::Bool=true) where T <: Real Description: Calculation of the spherical harmonic coefficients and expansion based on the measured t-design

Input:

tDesign - Measured t-design (type: SphericalTDesign)

field - Measured field (size = (J,N,C)) with J <= 3

x, y, z - Cartesian coordinates

kwargs:

L - Order up to which the coeffs be calculated (default: t/2)

calcSolid - Boolean (default: true)
false -> spherical coefficients
true -> solid coefficients

Output:

coeffs - spherical/solid coefficients, type: Array{SphericalHarmonicCoefficients}(3,C)

expansion - related expansion (Cartesian polynomial), type: Array{AbstractPolynomialLike}(3,C)

func - expansion converted to a function, type: Array{Function}(3,C)