MPISphericalHarmonics

Documentation for MPISphericalHarmonics.

Base.:+

Base.:-

MPISphericalHarmonics.findFFP

MPISphericalHarmonics.findFFP!

MPISphericalHarmonics.magneticField

MPISphericalHarmonics.shiftFFP!

Base.:+ — Method

+(mfc1::MagneticFieldCoefficients, mfc2::MagneticFieldCoefficients; force::Bool=false)

force = true adds the coefficients even if the radius or center are not equal (set to values of the first coefficients).

Base.:- — Method

-(mfc1::MagneticFieldCoefficients, mfc2::MagneticFieldCoefficients; force::Bool=false)

force = true subtracts the coefficients even if the radius or center are not equal (set to values of the first coefficients).

MPISphericalHarmonics.findFFP! — Method

ffp = findFFP!(coeffsMF::MagneticFieldCoefficients)

Find FFP and set it as coeffsMF.ffp.

MPISphericalHarmonics.findFFP — Method

findFFP(coeffsMF::MagneticFieldCoefficients; 
        returnasmatrix::Bool=true)

Description: Newton method to find the FFPs of the magnetic fields

Input:

coeffsMF - MagneticFieldCoefficients

kwargs:

returnasmatrix - Boolean
true -> return FFPs as Matrix with size (3,#Patches) (default)
false -> return FFPs as Array of NLsolve.SolverResults with size #Patches

Output:

ffp - FFPs of the magnetic field

MPISphericalHarmonics.magneticField — Method

magneticField(tDesign::SphericalTDesign, field::Union{AbstractArray{T,2},AbstractArray{T,3}};
	       L::Int=Int(tDesign.T/2),
	       calcSolid::Bool=true) where T <: Real

Description: Calculation of the spherical harmonic coefficients based on the measured t-design

Input:

tDesign - Measured t-design (type: SphericalTDesign)

field - Measured field (size = (J,N,C)) with J <= 3

kwargs:

L - Order up to which the coeffs be calculated (default: t/2)

calcSolid - Boolean (default: true)
false -> spherical coefficients
true -> solid coefficients

Output:

coeffs - spherical/solid coefficients, type: Array{SphericalHarmonicCoefficients}(3,C)

MPISphericalHarmonics.shiftFFP! — Method

shiftFFP!(coeffsMF::MagneticFieldCoefficients)

Shift magnetic-field coefficients into FFP (and calculate it if not available).